emacs/conf/mode-line: fix 'read-only buffer' error
This is all in context of running exwm.
Previously, checking for a single monitor to decide
whether to show system stats in mode-line caused 'read-only buffer'
error in gui buffers (such as browsers). This error would not allow
using native shortcuts on such buffers.
More context for future reference:
Apparently, when shell command is run (e.g., with
'shell-command-to-string), a temporary buffer is created to retrieve
the shell output. In this case, when switching to a buffer displaying
a browser, any attempt to operate on it with shortcuts would cause the
mentioned error.
The solution was to run shell processes with timer separate from
mode-line. These processes would right to a variable and mode-line
read these variable.