emacs/conf: fix creating monthly note
Now, traverse weeks in a month only by Mondays starting from the month
first day. (before it was the week day of the first month day. In that
case, days in the week beginning could ignored).

Always create monthly note from the 1st day of the month, even when
any date is selected by the user.