emacs/conf: comment out code for displaying MuPdf buffers
does not behave as wanted: The buffer only visible when opened from
workspaces where it is supposed to be displayed. If attempted to
display MuPdf while being in any other workspace, buffer is not
visible.

This behavior was already observed when displaying other buffers
that rely on X for displaying.