emacs/conf: try fixing setting then resetting keybindings
Some commands rebind keybindings in minibuffer-local-completion-map
on starting the completion and reset the old binding on
- exit
- error
- quit
However, when there was no original binding to the key,
a key with now binding will be left in the key-map.
This would cause "wrong argument type error" when other
commands try to modify the same key-map.