emacs/conf: fix exwm-workspace-switch bug
- when there are split windows in a frame, switching to that frame
  will always select the bottom window, regardless which window was
  selected before in the target frame.
- Also, when displaying a buffer which is supposed to be displayed in
  split window, will move focus to the bottom window as well.

Apparently, setting
- (setq exwm-workspace-warp-cursor t)
to move mouse cursor to the target workspace would
cause that unwanted behavior.
Resetting it back to default value
- (setq exwm-workspace-warp-cursor nil)
Would (temporarily) fix the problem.

Refer to my personal note on the matter to see the exact code in
- #'exwm-workspace-switch
where that happens (may be provide real fix for the problem).