emacs/conf: enable ram-toggle-lisp-editing-modes-in-org-code-block
The effects of this function are not perfect.
It is trying to do too many complex things.
And some edge cases are not covered.

I disabled it in the past because I though it was responsible for
incorrect #'org-mark-element behavior (It seem to mark correctly but
the yanked content includes extra lines at the end).

This functions did not cause this bug.
So, I enable it again and see how it behaves.