emacs/conf: modify capturing/creating weekly and montly notes
Check whether the target heading has subheadings before folding it.
This is done because the last heading in the tree will display "..."
as if it was folded, but there is no subheadings in it to be folded.
Is there a better way to avoid display a heading as if it was folded
when it has no content at all.