emac/conf: significant update to searches using minibuffer
Functions like those looking for:
  - a help on
    + function
    + variable
  - an outline
  - a definition
  - an Org #+name enty
  - etc
are updated.

This started with fixing an error
when search item was killed (in a kill ring). In some functions,
this would start debugging with a *Backtrace* buffer: the reason
being that the minibuffer could not handle 'quit minibuffer' command
nicely.

Also, some old issues were address, like:
  - incorrect search history handling.
  - adding the ability to change search commands on the fly,
    for example:
    - change search for and outline to searching for a #+name:
      while preserving the search string already typed in.
  - preserve the depth of
    + outlines
    + #+name: elements
    thus facilitating recognition and retention of their structure.